Moving kmeans and linear regression classes into a statistics package.
