+++ src/main/scala/com/microsoft/partnercatalyst/fortis/spark/statistics/ComputedAnomalyKmeansModelKey.scala
@@ -1,4 +1,6 @@
-package com.microsoft.partnercatalyst.fortis.spark.analyzer.timeseries
+package com.microsoft.partnercatalyst.fortis.spark.statistics
+
+import com.microsoft.partnercatalyst.fortis.spark.analyzer.timeseries.PeriodType
 
 object ComputedAnomalyKmeansModelKey {
